+++ src/httpserver/httphandlers/EpicsHttpHandler.java
@@ -1,4 +1,4 @@
-package httpServer.httphandlers;
+package httpserver.httphandlers;
 
 import com.google.gson.Gson;
 import com.google.gson.JsonSyntaxException;
