+++ src/httpserver/httphandlers/HistoryHttpHandler.java
@@ -1,4 +1,4 @@
-package httpServer.httphandlers;
+package httpserver.httphandlers;
 
 import com.google.gson.Gson;
 import com.sun.net.httpserver.HttpExchange;
